Move CalcJob base class to calculations/base.py
