+++ aiida_quantumespresso/calculations/neb.py
@@ -9,11 +9,12 @@ import six
 from aiida import orm
 from aiida.common import InputValidationError, CalcInfo, CodeInfo
 from aiida.common.lang import classproperty
-from aiida_quantumespresso._base import CalcJob
 
 from aiida_quantumespresso.calculations.pw import PwCalculation
 from aiida_quantumespresso.calculations import _lowercase_dict, _uppercase_dict, _pop_parser_options
 from aiida_quantumespresso.utils.convert import convert_input_to_namelist_entry
+
+from .base import CalcJob
 
 
 class NebCalculation(CalcJob):
