+++ aiida_quantumespresso/calculations/ph.py
@@ -9,10 +9,11 @@ import six
 from aiida import orm
 from aiida.common import datastructures, exceptions
 
-from aiida_quantumespresso._base import CalcJob
 from aiida_quantumespresso.calculations.pw import PwCalculation
 from aiida_quantumespresso.calculations import _lowercase_dict, _uppercase_dict
 from aiida_quantumespresso.utils.convert import convert_input_to_namelist_entry
+
+from .base import CalcJob
 
 
 class PhCalculation(CalcJob):
